+++ packages/react/src/components/HotspotEditorModal/HotspotEditorModal.jsx
@@ -295,6 +295,7 @@ const HotspotEditorModal = ({
     },
   });
 
+  const hasNonEditableContent = React.isValidElement(selectedHotspot?.content);
   const mergedI18n = useMemo(() => ({ ...defaultProps.i18n, ...i18n }), [i18n]);
 
   const {
@@ -493,7 +494,9 @@ const HotspotEditorModal = ({
                 translateWithId={translateWithId}
               />
             </Tab>
-            <Tab label={fixedTypeDataSourceTabLabelText}>{renderDataSourceTab()}</Tab>
+            <Tab disabled={hasNonEditableContent} label={fixedTypeDataSourceTabLabelText}>
+              {hasNonEditableContent ? null : renderDataSourceTab()}
+            </Tab>
           </Tabs>
         )}
       </>
@@ -538,7 +541,9 @@ const HotspotEditorModal = ({
             translateWithId={translateWithId}
           />
         </Tab>
-        <Tab label={textTypeDataSourceTabLabelText}>{renderDataSourceTab()}</Tab>
+        <Tab disabled={hasNonEditableContent} label={textTypeDataSourceTabLabelText}>
+          {hasNonEditableContent ? null : renderDataSourceTab()}
+        </Tab>
       </Tabs>
     );
   };