+++ packages/react/src/components/HotspotEditorModal/HotspotEditorModal.story.jsx
@@ -156,6 +156,19 @@ export const WithExistingHotspots = withReadme(HotspotEditorModalREADME, () => {
       ...cardConfig.values,
       hotspots: [
         {
+          x: 45,
+          y: 25,
+          color: green50,
+          content: <span style={{ padding: '1rem' }}>content is an element</span>,
+        },
+        {
+          x: 65,
+          y: 75,
+          type: 'text',
+          color: green50,
+          content: <span>content is an element</span>,
+        },
+        {
           x: 75,
           y: 10,
           type: 'text',