+++ packages/react/src/components/HotspotEditorModal/HotspotEditorTooltipTab/HotspotEditorTooltipTab.jsx
@@ -142,6 +142,7 @@ const HotspotEditorTooltipTab = ({
   } = merge({}, defaultProps.i18n, i18n);
 
   const currentIconColor = formValues.color?.carbonColor ?? formValues.color ?? 'currentcolor';
+  const hasNonEditableContent = React.isValidElement(formValues?.content);
 
   const renderInfoMessage = () => (
     <div className={`${iotPrefix}--hotspot-editor--tooltip-info-message`}>
@@ -201,6 +202,7 @@ const HotspotEditorTooltipTab = ({
             onSubmit={preventFormSubmission}
           >
             <MyTitleTextInput
+              disabled={hasNonEditableContent}
               name="title"
               data-testid={`${testID}-title-input`}
               value={formValues.content?.title || ''}
@@ -215,6 +217,7 @@ const HotspotEditorTooltipTab = ({
               {...getOverrides(overrides?.titleTextInput?.props)}
             />
             <MyDecriptionTextArea
+              disabled={hasNonEditableContent}
               name="description"
               id="tooltip-form-description"
               labelText={descriptionTextareaLabelText}